requirements.txt
Django==2.2.10
djangorestframework==3.12.4
django-cors-headers==3.7.0
drf_yasg==1.20.0
psycopg2_binary>=2.8,<2.9
python_decouple==3.4